refactor: index.js 생성하여 컴포넌트 파일 정리
