+++ front/src/components/index.js
@@ -1,0 +1,4 @@
+export { default as Button } from "./Button";
+export { default as MemoCard } from "./MemoCard";
+export { default as MemoModal } from "./MemoModal";
+export { default as SearchBar } from "./SearchBar";
