+++ front/src/pages/LoginPage.jsx
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import Button from "../components/Button";
+import { Button } from "../components";
 import { Link } from "react-router-dom";
 
 const LoginPageContainer = styled.div`
