+++ front/src/pages/SignUpPage.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import styled from "@emotion/styled";
-import Button from "../components/Button";
+import { Button } from "../components";
 import { Link } from "react-router-dom";
 
 const SignUpPageContainer = styled.div`
